[Silver IV] Title: 스케이트 연습, Time: 184 ms, Memory: 51040 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/28324. 스케이트 연습/스케이트 연습.cs
@@ -4,8 +4,8 @@
 int n = int.Parse(Console.ReadLine());
 int[] inputVelo = Array.ConvertAll(Console.ReadLine().Split(), int.Parse);
 
-int speed = 0;
-int ans = 0;
+long speed = 0;
+long ans = 0;
 
 for (int i = n - 1; i >= 0; i--)
 {
